added explicit cast to DWORD to prevent unsigned/signed warning on some MSVC compile settings.

+++ src/io/interfaces/windows/WindowsDeviceInterface.cpp
@@ -81,7 +81,7 @@ namespace zen
                 return ZenError_Io_SendFailed;
         }
 
-        if (nBytesWritten != data.size())
+        if (nBytesWritten != static_cast<DWORD>(data.size()))
             return ZenError_Io_SendFailed;
 
         return ZenError_None;
